+++ src/fdw.c
@@ -1994,8 +1994,8 @@ is_simple_join_clause(Expr* expr) {
     if (IsA(expr, OpExpr)) {
         OpExpr* opexpr = (OpExpr*)expr;
 
-        if (chfdw_is_equal_op(opexpr->opno) == 1 && list_length(opexpr->args) == 2 &&
-            IsA(list_nth(opexpr->args, 0), Var) &&
+        if (chfdw_is_equal_op(opexpr->opno) == CH_OP_EQ &&
+            list_length(opexpr->args) == 2 && IsA(list_nth(opexpr->args, 0), Var) &&
             IsA(list_nth(opexpr->args, 1), Var)) {
             return true;
         }